Updated  Unity, Created New Enemies, Expanded Dialog System
I did a looooot of refactoring of enemy scripts, then used those to actually try to implement some enemies I thought were interesting. Included in this was creating a boss, which I'd already conceptualized and made in game, but I threw out most of the implemented code for it and used the newly refactored enemy stuff instead. I also went back to my dialog parsing code and made a lot of changes. In particular, I put some more focus on accessing the global dictionary when calculating dialog, and also modified the parser so that nested if's work, something I had previously neglected.